dev-requirements.txt
-r requirements.txt
coverage>=4.4.2
flake8>=3.5.0
hypothesis>=3.44.3
pytest>=3.3.0
pytest-cov>=2.5.1
pytest-django>=3.1.2
